Implement alias management functionality and refactor app.js
- Introduced a new `aliasView.js` module to handle alias grid rendering and management, enhancing code organization and maintainability.
- Updated `app.js` to integrate the new alias management functionality, improving separation of concerns within the application.
- Refactored existing alias-related methods to utilize the new module, streamlining the code and enhancing readability.
- Added comprehensive tests to validate the new alias functionality and ensure UI elements reflect changes accurately.

These updates significantly enhance the application's alias management capabilities and overall user interface responsiveness, contributing to a more intuitive user experience.
